+++ bee-ledger/src/workers/consensus/worker.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    types::{Balance, CreatedOutput, LedgerIndex, Migration, Receipt, TreasuryOutput},
+    types::{CreatedOutput, LedgerIndex, Migration, Receipt, TreasuryOutput},
     workers::{
         consensus::{metadata::WhiteFlagMetadata, state::validate_ledger_state, white_flag},
         error::Error,
@@ -39,8 +39,6 @@ pub(crate) const EXTRA_PRUNING_DEPTH: u32 = 5;
 pub enum ConsensusWorkerCommand {
     /// Command to confirm a milestone.
     ConfirmMilestone(MessageId),
-    /// Command to fetch the balance of an address.
-    FetchBalance(Address, oneshot::Sender<(Result<Option<Balance>, Error>, LedgerIndex)>),
     /// Command to fetch an output.
     FetchOutput(
         OutputId,
@@ -145,7 +143,6 @@ where
                     Output::from(ExtendedOutput::new(*fund.address(), fund.amount())),
                 ),
             );
-            metadata.balance_diffs.amount_add(*fund.address(), fund.amount())?;
         }
 
         if receipt.migrated_at() < *receipt_migrated_at {
@@ -179,7 +176,6 @@ where
         metadata.milestone_index,
         &metadata.created_outputs,
         &metadata.consumed_outputs,
-        &metadata.balance_diffs,
         &migration,
     )?;
 
@@ -368,11 +364,6 @@ where
                             Err(reason) => {
                                 debug!("Pruning skipped: {:?}", reason);
                             }
-                        }
-                    }
-                    ConsensusWorkerCommand::FetchBalance(address, sender) => {
-                        if let Err(e) = sender.send((storage::fetch_balance(&*storage, &address), ledger_index)) {
-                            error!("Error while sending balance: {:?}", e);
                         }
                     }
                     ConsensusWorkerCommand::FetchOutput(output_id, sender) => {