+++ bee-storage/bee-storage-memory/src/access/truncate.rs
@@ -6,11 +6,10 @@
 use crate::storage::Storage;
 
 use bee_ledger::types::{
-    snapshot::SnapshotInfo, Balance, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt, TreasuryOutput,
-    Unspent,
+    snapshot::SnapshotInfo, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt, TreasuryOutput, Unspent,
 };
 use bee_message::{
-    address::{Address, Ed25519Address},
+    address::Ed25519Address,
     milestone::{Milestone, MilestoneIndex},
     output::OutputId,
     payload::indexation::PaddedIndex,
@@ -46,7 +45,6 @@ impl_truncate!(MilestoneIndex, Milestone, milestone_index_to_milestone);
 impl_truncate!((), SnapshotInfo, snapshot_info);
 impl_truncate!(SolidEntryPoint, MilestoneIndex, solid_entry_point_to_milestone_index);
 impl_truncate!(MilestoneIndex, OutputDiff, milestone_index_to_output_diff);
-impl_truncate!(Address, Balance, address_to_balance);
 impl_truncate!(
     (MilestoneIndex, UnreferencedMessage),
     (),