+++ bee-storage/bee-storage-rocksdb/src/access/batch.rs
@@ -7,11 +7,11 @@ use crate::{
 };
 
 use bee_ledger::types::{
-    snapshot::info::SnapshotInfo, Balance, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt,
-    TreasuryOutput, Unspent,
+    snapshot::info::SnapshotInfo, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt, TreasuryOutput,
+    Unspent,
 };
 use bee_message::{
-    address::{Address, Ed25519Address},
+    address::Ed25519Address,
     milestone::{Milestone, MilestoneIndex},
     output::OutputId,
     payload::indexation::PaddedIndex,
@@ -479,31 +479,6 @@ impl Batch<MilestoneIndex, OutputDiff> for Storage {
         batch
             .inner
             .delete_cf(self.cf_handle(CF_MILESTONE_INDEX_TO_OUTPUT_DIFF)?, &batch.key_buf);
-
-        Ok(())
-    }
-}
-
-impl Batch<Address, Balance> for Storage {
-    fn batch_insert(
-        &self,
-        batch: &mut Self::Batch,
-        address: &Address,
-        balance: &Balance,
-    ) -> Result<(), <Self as StorageBackend>::Error> {
-        batch.inner.put_cf(
-            self.cf_handle(CF_ADDRESS_TO_BALANCE)?,
-            address.pack_to_vec(),
-            balance.pack_to_vec(),
-        );
-
-        Ok(())
-    }
-
-    fn batch_delete(&self, batch: &mut Self::Batch, address: &Address) -> Result<(), <Self as StorageBackend>::Error> {
-        batch
-            .inner
-            .delete_cf(self.cf_handle(CF_ADDRESS_TO_BALANCE)?, address.pack_to_vec());
 
         Ok(())
     }