+++ bee-storage/bee-storage-rocksdb/src/access/delete.rs
@@ -7,11 +7,11 @@ use crate::{
 };
 
 use bee_ledger::types::{
-    snapshot::info::SnapshotInfo, Balance, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt,
-    TreasuryOutput, Unspent,
+    snapshot::info::SnapshotInfo, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt, TreasuryOutput,
+    Unspent,
 };
 use bee_message::{
-    address::{Address, Ed25519Address},
+    address::Ed25519Address,
     milestone::{Milestone, MilestoneIndex},
     output::OutputId,
     payload::indexation::PaddedIndex,
@@ -144,15 +144,6 @@ impl Delete<MilestoneIndex, OutputDiff> for Storage {
     fn delete(&self, index: &MilestoneIndex) -> Result<(), <Self as StorageBackend>::Error> {
         self.inner
             .delete_cf(self.cf_handle(CF_MILESTONE_INDEX_TO_OUTPUT_DIFF)?, index.pack_to_vec())?;
-
-        Ok(())
-    }
-}
-
-impl Delete<Address, Balance> for Storage {
-    fn delete(&self, address: &Address) -> Result<(), <Self as StorageBackend>::Error> {
-        self.inner
-            .delete_cf(self.cf_handle(CF_ADDRESS_TO_BALANCE)?, address.pack_to_vec())?;
 
         Ok(())
     }