+++ bee-storage/bee-storage-rocksdb/src/access/exist.rs
@@ -7,11 +7,11 @@ use crate::{
 };
 
 use bee_ledger::types::{
-    snapshot::info::SnapshotInfo, Balance, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt,
-    TreasuryOutput, Unspent,
+    snapshot::info::SnapshotInfo, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt, TreasuryOutput,
+    Unspent,
 };
 use bee_message::{
-    address::{Address, Ed25519Address},
+    address::Ed25519Address,
     milestone::{Milestone, MilestoneIndex},
     output::OutputId,
     payload::indexation::PaddedIndex,
@@ -152,15 +152,6 @@ impl Exist<MilestoneIndex, OutputDiff> for Storage {
         Ok(self
             .inner
             .get_cf(self.cf_handle(CF_MILESTONE_INDEX_TO_OUTPUT_DIFF)?, index.pack_to_vec())?
-            .is_some())
-    }
-}
-
-impl Exist<Address, Balance> for Storage {
-    fn exist(&self, address: &Address) -> Result<bool, <Self as StorageBackend>::Error> {
-        Ok(self
-            .inner
-            .get_cf(self.cf_handle(CF_ADDRESS_TO_BALANCE)?, address.pack_to_vec())?
             .is_some())
     }
 }