+++ bee-storage/bee-storage-rocksdb/src/access/insert.rs
@@ -7,11 +7,11 @@ use crate::{
 };
 
 use bee_ledger::types::{
-    snapshot::info::SnapshotInfo, Balance, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt,
-    TreasuryOutput, Unspent,
+    snapshot::info::SnapshotInfo, ConsumedOutput, CreatedOutput, LedgerIndex, OutputDiff, Receipt, TreasuryOutput,
+    Unspent,
 };
 use bee_message::{
-    address::{Address, Ed25519Address},
+    address::Ed25519Address,
     milestone::{Milestone, MilestoneIndex},
     output::OutputId,
     payload::indexation::PaddedIndex,
@@ -185,18 +185,6 @@ impl Insert<MilestoneIndex, OutputDiff> for Storage {
             self.cf_handle(CF_MILESTONE_INDEX_TO_OUTPUT_DIFF)?,
             index.pack_to_vec(),
             diff.pack_to_vec(),
-        )?;
-
-        Ok(())
-    }
-}
-
-impl Insert<Address, Balance> for Storage {
-    fn insert(&self, address: &Address, balance: &Balance) -> Result<(), <Self as StorageBackend>::Error> {
-        self.inner.put_cf(
-            self.cf_handle(CF_ADDRESS_TO_BALANCE)?,
-            address.pack_to_vec(),
-            balance.pack_to_vec(),
         )?;
 
         Ok(())