+++ bee-storage/bee-storage-rocksdb/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 0.7.0 - 2022-XX-XX
+
+### Removed
+
+- All `Balance` operations;
+
 ## 0.6.0 - 2022-XX-XX
 
 ### Changed